nitrates(front): #284 — intègre l'animation groovy au popup + retire les sandboxes
L'animation « vers groovy » validée devient l'animation reward officielle :
feedback_vers.js/css portent désormais la version horizontale (traversée +
croisements), la fonction garde le nom nitratesVersReward (déjà câblé dans le
popup). Suppression des fichiers WIP feedback_vers_groovy.*, des deux pages
sandbox (/sandbox/vers/ et /sandbox/vers-design/) et de leurs routes DEBUG.
