+++ envergo/static/nitrates/feedback_vers.css
@@ -1,60 +1,97 @@
-/* #284 — Animation "reward" de fin de feedback : des petits lombrics contents
-   qui flottent et sourient. Couleur ver de terre (rosé-brun), sobre mais
-   sympa (on est l'État, mais on a le droit de faire plaisir).
+/* #284 — WIP animation "vers groovy" : ils traversent horizontalement (des uns
+   vers la droite, des autres vers la gauche -> ils se croisent), en ondulant
+   façon danse. Page de design dédiée, on itère avant intégration. */
 
-   Structure : un conteneur .nitrates-vers qui overlay le succès du formulaire.
-   Chaque .nitrates-ver est un SVG inline positionné, qui "flotte" (bob) et
-   apparaît avec un petit zoom-in décalé (stagger via --delai). */
-
-.nitrates-vers {
+.vg-scene {
   position: relative;
   width: 100%;
-  min-height: 140px;
-  pointer-events: none;
-  overflow: visible;
+  height: 160px;
+  overflow: hidden;
 }
 
-.nitrates-ver {
+.vg-ver {
   position: absolute;
-  width: 64px;
-  height: 64px;
-  transform-origin: center bottom;
-  opacity: 0;
+  width: 120px;
+  height: 48px;
+  will-change: transform;
+}
+
+/* Déplacement horizontal complet (de hors-champ gauche à hors-champ droite) +
+   petit dandinement vertical (bob) qui donne le côté groovy. */
+.vg-vers-droite {
+  left: 0;
   animation:
-    nitrates-ver-pop 0.5s cubic-bezier(0.34, 1.56, 0.64, 1) forwards,
-    nitrates-ver-float 3.2s ease-in-out infinite;
-  animation-delay: var(--delai, 0s), calc(var(--delai, 0s) + 0.5s);
+    vg-traverse-droite var(--dur, 6s) linear var(--delai, 0s) infinite,
+    vg-bob 0.9s ease-in-out infinite;
 }
 
-/* Apparition : zoom-in rebondi. */
-@keyframes nitrates-ver-pop {
-  0% {
-    opacity: 0;
-    transform: translateY(12px) scale(0.2) rotate(var(--rot, 0deg));
-  }
+.vg-vers-gauche {
+  left: 0;
+  /* miroir horizontal : la tête (dessinée à droite) pointe vers la gauche */
+  animation:
+    vg-traverse-gauche var(--dur, 6s) linear var(--delai, 0s) infinite,
+    vg-bob 0.9s ease-in-out infinite;
+}
 
-  100% {
-    opacity: 1;
-    transform: translateY(0) scale(1) rotate(var(--rot, 0deg));
+.vg-vers-gauche .vg-svg {
+  transform: scaleX(-1);
+}
+
+/* Distance parcourue = --trajet (largeur RÉELLE de la scène + un ver), calculée
+   en pixels par le JS (le `%` d'un translateX réfère à l'élément, pas au
+   conteneur -> il fallait la vraie largeur). Le ver part hors-champ (-120px) et
+   sort à droite (--trajet). Les deux sens utilisent la MÊME distance -> une même
+   `--dur` = une même vitesse quel que soit le sens. */
+@keyframes vg-traverse-droite {
+  from {
+    transform: translateX(-120px);
+  }
+  to {
+    transform: translateX(var(--trajet, 100%));
   }
 }
 
-/* Flottement doux + léger balancement, indépendant par ver via --amp/--dur. */
-@keyframes nitrates-ver-float {
+@keyframes vg-traverse-gauche {
+  from {
+    transform: translateX(var(--trajet, 100%));
+  }
+  to {
+    transform: translateX(-120px);
+  }
+}
+
+/* Dandinement vertical : petit bob + inclinaison, façon groove. */
+@keyframes vg-bob {
   0%,
   100% {
-    transform: translateY(0) rotate(var(--rot, 0deg));
+    margin-top: 0;
   }
-
   50% {
-    transform: translateY(-10px) rotate(calc(var(--rot, 0deg) + 4deg));
+    margin-top: -8px;
   }
 }
 
-/* Respect de prefers-reduced-motion : pas de flottement, juste l'apparition. */
+/* Ondulation "reptation" du corps : on fait légèrement pulser l'échelle
+   verticale du corps pour simuler la vague qui parcourt le ver. */
+.vg-corps {
+  animation: vg-onde 0.7s ease-in-out infinite;
+  transform-origin: center;
+}
+
+@keyframes vg-onde {
+  0%,
+  100% {
+    transform: scaleY(1);
+  }
+  50% {
+    transform: scaleY(0.86);
+  }
+}
+
 @media (prefers-reduced-motion: reduce) {
-  .nitrates-ver {
-    animation: nitrates-ver-pop 0.4s ease forwards;
-    animation-delay: var(--delai, 0s);
+  .vg-vers-droite,
+  .vg-vers-gauche,
+  .vg-corps {
+    animation-duration: 0s;
   }
 }
