--- envergo/templates/nitrates/sandbox_vers.html
@@ -1,79 +1,0 @@
-{% extends "nitrates/base.html" %}
-
-{% load static %}
-
-{% comment %}
-SANDBOX #284 (temporaire) : page de test de l'animation "reward" de fin de
-feedback (les petits lombrics contents). Permet de trigger l'animation à la
-demande pour la valider visuellement. À SUPPRIMER une fois le dev fini
-(template + URL + vue). N'est câblée qu'en DEBUG.
-{% endcomment %}
-
-{% block title %}Sandbox vers{% endblock %}
-
-{% block extra_css %}
-  <link href="{% static 'nitrates/feedback_vers.css' %}" rel="stylesheet">
-  <style>
-    .sandbox-wrap {
-      max-width: 720px;
-      margin: 3rem auto;
-      text-align: center;
-    }
-
-    .sandbox-stage {
-      margin: 2rem auto;
-      padding: 2rem;
-      border: 1px dashed #ccc;
-      border-radius: 8px;
-      min-height: 200px;
-      display: flex;
-      align-items: center;
-      justify-content: center;
-    }
-
-    .sandbox-stage__inner {
-      width: 360px;
-    }
-  </style>
-{% endblock %}
-
-{% block content %}
-  <div class="fr-container sandbox-wrap">
-    <h1 class="fr-h3">Sandbox — animation « vers contents » (#284)</h1>
-    <p class="fr-text--sm">Page de test temporaire. Clique pour relancer l'animation.</p>
-
-    <div class="fr-btns-group fr-btns-group--inline fr-btns-group--center">
-      <button type="button" class="fr-btn" id="sandbox-rejouer">Rejouer l'animation</button>
-      <button type="button" class="fr-btn fr-btn--secondary" id="sandbox-moins">– 1 ver</button>
-      <button type="button" class="fr-btn fr-btn--secondary" id="sandbox-plus">+ 1 ver</button>
-    </div>
-    <p class="fr-text--sm fr-mt-1w">
-      Nombre de vers : <strong id="sandbox-nb">5</strong>
-    </p>
-
-    <div class="sandbox-stage">
-      <div class="sandbox-stage__inner" id="sandbox-stage"></div>
-    </div>
-  </div>
-
-  <script src="{% static 'nitrates/feedback_vers.js' %}"></script>
-  <script>
-    (function () {
-      var nb = 5;
-      var stage = document.getElementById("sandbox-stage");
-      var labelNb = document.getElementById("sandbox-nb");
-      function rejouer() {
-        labelNb.textContent = nb;
-        window.nitratesVersReward(stage, { nombre: nb });
-      }
-      document.getElementById("sandbox-rejouer").addEventListener("click", rejouer);
-      document.getElementById("sandbox-plus").addEventListener("click", function () {
-        nb = Math.min(5, nb + 1); rejouer();
-      });
-      document.getElementById("sandbox-moins").addEventListener("click", function () {
-        nb = Math.max(1, nb - 1); rejouer();
-      });
-      rejouer();
-    })();
-  </script>
-{% endblock %}
