--- envergo/templates/nitrates/sandbox_vers_design.html
@@ -1,76 +1,0 @@
-{% extends "nitrates/base.html" %}
-{% load static %}
-
-{% comment %}
-SANDBOX DESIGN #284 (temporaire) : page dédiée au design des vers de terre
-"groovy" qui traversent horizontalement et se croisent. On itère ici avec Max
-avant d'intégrer dans le popup de feedback. À SUPPRIMER en fin de dev (template
-+ URL + feedback_vers_groovy.*). DEBUG uniquement.
-{% endcomment %}
-
-{% block title %}Design vers groovy{% endblock %}
-
-{% block extra_css %}
-  <link href="{% static 'nitrates/feedback_vers_groovy.css' %}" rel="stylesheet">
-  <style>
-    .vgd-wrap {
-      max-width: 900px;
-      margin: 2.5rem auto;
-      text-align: center;
-    }
-
-    .vgd-stage {
-      margin: 1.5rem auto;
-      border: 1px dashed #ccc;
-      border-radius: 8px;
-      background: #fff;
-    }
-
-    .vgd-single {
-      margin: 2rem auto;
-      width: 160px;
-    }
-  </style>
-{% endblock %}
-
-{% block content %}
-  <div class="fr-container vgd-wrap">
-    <h1 class="fr-h4">Design — vers de terre « groovy » (#284)</h1>
-    <p class="fr-text--sm">
-      Page de test temporaire. Ils traversent, se croisent, sourient. On itère
-      ici avant d'intégrer au popup de feedback.
-    </p>
-
-    <div class="fr-btns-group fr-btns-group--inline fr-btns-group--center">
-      <button type="button" class="fr-btn" id="vgd-rejouer">Rejouer</button>
-    </div>
-
-    {# Scène animée : les vers traversent et se croisent. #}
-    <div class="vgd-stage" id="vgd-stage"></div>
-
-    {# Un ver figé, en grand, pour juger le design du personnage seul. #}
-    <p class="fr-text--sm fr-mt-2w">Le personnage seul (figé, agrandi) :</p>
-    <div class="vgd-single" id="vgd-single"></div>
-  </div>
-
-  <script src="{% static 'nitrates/feedback_vers_groovy.js' %}"></script>
-  <script>
-    (function () {
-      var stage = document.getElementById("vgd-stage");
-      var single = document.getElementById("vgd-single");
-      function rejouer() {
-        window.nitratesVersGroovy(stage, { nombre: 4 });
-        // Personnage seul, figé et agrandi (pas d'animation de traversée) :
-        single.innerHTML =
-          '<div class="vg-ver" style="position:static;width:160px;height:64px">' +
-          '<svg viewBox="0 0 120 48" width="160" height="64"></svg></div>';
-        // On réutilise la même fonction de dessin via un ver monté puis extrait.
-        var tmp = document.createElement("div");
-        window.nitratesVersGroovy(tmp, { nombre: 1 });
-        single.innerHTML = tmp.querySelector(".vg-ver").innerHTML;
-      }
-      document.getElementById("vgd-rejouer").addEventListener("click", rejouer);
-      rejouer();
-    })();
-  </script>
-{% endblock %}
